+++ EAGRN-JMetal/src/main/java/eagrn/WeightedConfRunner.java
@@ -34,13 +34,13 @@ public class WeightedConfRunner {
         }
 
         /** Read all files with lists. */
-        Map<String, Double[]> inferredNetworks = StaticUtils.readAllInferredNetworkFiles(inferredNetworkFiles);
+        Map<String, Float[]> inferredNetworks = StaticUtils.readAllInferredNetworkFiles(inferredNetworkFiles);
 
         /** Calculate the list of weighted confidence levels based on the vector of weights. */
-        Map<String, Double> weightedConf = StaticUtils.makeConsensus(weights, inferredNetworks);
+        Map<String, Float> weightedConf = StaticUtils.makeConsensus(weights, inferredNetworks);
 
         /** Sort map */
-        Map<String, Double> weightedConfSort = weightedConf
+        Map<String, Float> weightedConfSort = weightedConf
                 .entrySet()
                 .stream()
                 .sorted(Collections.reverseOrder(Map.Entry.comparingByValue()))
