+++ EAGRN-JMetal/src/main/java/eagrn/fitnessfunction/FitnessFunction.java
@@ -3,5 +3,5 @@ package eagrn.fitnessfunction;
 import java.util.Map;
 
 public interface FitnessFunction {
-    public double run(Map<String, Double> consensus, Double[] x);
+    public double run(Map<String, Float> consensus, Double[] x);
 }
