+++ EAGRN-JMetal/src/main/java/eagrn/fitnessfunction/impl/quality/impl/QualityMeanAboveCutOff.java
@@ -18,28 +18,28 @@ import java.util.Map;
 public class QualityMeanAboveCutOff extends Quality {
     private CutOffCriteria cutOffCriteria;
     
-    public QualityMeanAboveCutOff(Map<String, Double[]> inferredNetworks, CutOffCriteria cutOffCriteria) {
+    public QualityMeanAboveCutOff(Map<String, Float[]> inferredNetworks, CutOffCriteria cutOffCriteria) {
         super(inferredNetworks, TrendMeasureEnum.MEAN);
         this.cutOffCriteria = cutOffCriteria;
     }
 
-    public double run(Map<String, Double> consensus, Double[] x) {
+    public double run(Map<String, Float> consensus, Double[] x) {
         /** 
          * 1. We create the map with the distances corresponding to the difference between the maximum and minimum 
          * of the vectors containing the mean between the weight and the distance normalized to the mean of the total 
          * number of techniques 
          */
-        Map<String, Double> distances = super.makeDistanceMap(consensus, x);
+        Map<String, Float> distances = super.makeDistanceMap(consensus, x);
 
         /** 2. Get high quality links */
-        Map<String, Double> goodLinks = cutOffCriteria.getCutMap(consensus);
+        Map<String, Float> goodLinks = cutOffCriteria.getCutMap(consensus);
 
         /** 3. Calculate the average of confidence-distance means of high quality links */
-        double conf, dist, confDistSum = 0;
-        for (Map.Entry<String, Double> pair : goodLinks.entrySet()) {
+        float conf, dist, confDistSum = 0;
+        for (Map.Entry<String, Float> pair : goodLinks.entrySet()) {
             conf = pair.getValue();
             dist = distances.get(pair.getKey());
-            confDistSum += (conf + (1 - dist)) / 2.0;
+            confDistSum += (float) ((conf + (1 - dist)) / 2.0);
         }
         double mean = confDistSum / goodLinks.size();
