+++ EAGRN-JMetal/src/test/java/eagrn/fitnessfunctions/impl/topology/impl/BinarizedDegreeDistributionTest.java
@@ -48,11 +48,11 @@ public class BinarizedDegreeDistributionTest {
         };
     }
 
-    private Map<String, Double> getMapFromGraph(Graph<String, DefaultEdge> graph){
-        Map<String, Double> map = new HashMap<>();
+    private Map<String, Float> getMapFromGraph(Graph<String, DefaultEdge> graph){
+        Map<String, Float> map = new HashMap<>();
         for (DefaultEdge edge : graph.edgeSet()) {
             String[] strEdge = edge.toString().split("[ ():]");
-            map.put(strEdge[1] + ";" + strEdge[4], 1.0);
+            map.put(strEdge[1] + ";" + strEdge[4], 1.0f);
         }
         return map;
     }
@@ -62,12 +62,12 @@ public class BinarizedDegreeDistributionTest {
         Graph<String, DefaultEdge> randomGraph = new SimpleDirectedGraph<>(getVertexSupplier(), SupplierUtil.createDefaultEdgeSupplier(), false);
         GnpRandomGraphGenerator<String, DefaultEdge> randomGenerator = new GnpRandomGraphGenerator<>(size, 0.5);
         randomGenerator.generateGraph(randomGraph);
-        Map<String, Double> badTopologyConsensus = getMapFromGraph(randomGraph);
+        Map<String, Float> badTopologyConsensus = getMapFromGraph(randomGraph);
 
         Graph<String, DefaultEdge> scaleFreeGraph = new SimpleDirectedGraph<>(getVertexSupplier(), SupplierUtil.createDefaultEdgeSupplier(), false);
         ScaleFreeGraphGenerator<String, DefaultEdge> scaleFreeGenerator = new ScaleFreeGraphGenerator<>(size);
         scaleFreeGenerator.generateGraph(scaleFreeGraph);
-        Map<String, Double> goodTopologyConsensus = getMapFromGraph(scaleFreeGraph);
+        Map<String, Float> goodTopologyConsensus = getMapFromGraph(scaleFreeGraph);
 
         Supplier<String> vSupplier = getVertexSupplier();
         ArrayList<String> geneNames = new ArrayList<>();
@@ -77,7 +77,7 @@ public class BinarizedDegreeDistributionTest {
 
         Double[] x = new Double[]{};
 
-        BinarizedDegreeDistribution topology = new BinarizedDegreeDistribution(size, new MinConfCriteria(0.5, geneNames));
+        BinarizedDegreeDistribution topology = new BinarizedDegreeDistribution(size, new MinConfCriteria(0.5f, geneNames));
         double fitnessGoodTopologyConsensus = topology.run(goodTopologyConsensus, x);
         double fitnessBadTopologyConsensus = topology.run(badTopologyConsensus, x);
